image_dashboard2/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
argcomplete==3.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncer==0.0.2
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
boto3==1.35.76
botocore==1.35.76
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
chainlit==0.7.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dacite==1.7.0
dataclasses-json==0.5.14
deap==1.4.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
distlib==0.3.9
dj-database-url==1.2.0
Django==5.0.6
django-dbbackup==4.1.0
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dnspython==2.7.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.100.1
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
ffmpy==0.4.0
fiftyone==1.1.0
fiftyone-brain==0.18.1
fiftyone_db==1.1.7
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
future==1.0.0
gast==0.6.0
git-filter-repo==2.38.0
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
google-pasta==0.2.0
googleapis-common-protos==1.66.0
gradio==5.8.0
gradio_client==1.5.1
graphql-core==3.2.5
grpcio==1.66.2
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.5
humanize==4.11.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
imageio==2.35.1
importlib_metadata==8.5.0
inflate64==1.0.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kaleido==0.2.1
keras==3.6.0
kiwisolver==1.4.5
langchain==0.0.305
langsmith==0.0.92
Lazify==0.4.0
lazy_loader==0.4
libclang==18.1.1
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mongoengine==0.24.2
motor==3.5.3
mpmath==1.3.0
multidict==6.1.0
multivolumefile==0.2.3
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.9.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
notebook==7.1.3
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.12
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pinecone-client==2.2.4
platformdirs==4.2.1
plotly==5.24.1
pprintpp==0.4.0
priority==2.0.0
prisma==0.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==5.29.1
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.22.0
pyarrow==17.0.0
pybcj==1.0.2
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pydash==8.0.4
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
pyglet==2.0.18
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
pyparsing==3.1.2
pyppmd==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.10.1
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-multipart==0.0.6
python-socketio==5.11.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
pyzstd==0.16.2
rarfile==4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.18.0
Rtree==1.3.0
ruff==0.8.2
s3transfer==0.10.4
safehttpx==0.1.6
scikit-fuzzy==0.5.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.5.0
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.27.0
strawberry-graphql==0.253.1
streamlit==1.39.0
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.8.30
tinycss2==1.3.0
tinydb==4.8.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ultralytics==8.3.47
ultralytics-thop==2.0.12
universal-analytics-python3==1.1.1
uptrace==1.28.2
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.23.2
virtualenv==20.28.0
voxel51-eta==0.13.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.4
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.18.3
zipp==3.21.0
